make code more c89 compatible

+++ src/snprintf.c
@@ -950,7 +950,7 @@ int portable_vsnprintf(char *str, size_t str_m, const char *fmt, va_list ap) {
         if (n > 0) {
           if (str_l < str_m) {
             size_t avail = str_m-str_l;
-            fast_memset(str+str_l, (zero_padding?'0':' '), (n>avail?avail:n));
+            fast_memset(str+str_l, (zero_padding?'0':' '), ((size_t)n>avail?avail:n));
           }
           str_l += n;
         }
@@ -967,7 +967,7 @@ int portable_vsnprintf(char *str, size_t str_m, const char *fmt, va_list ap) {
         if (n > 0) {
           if (str_l < str_m) {
             size_t avail = str_m-str_l;
-            fast_memcpy(str+str_l, str_arg, (n>avail?avail:n));
+            fast_memcpy(str+str_l, str_arg, ((size_t)n>avail?avail:n));
           }
           str_l += n;
         }
@@ -976,7 +976,7 @@ int portable_vsnprintf(char *str, size_t str_m, const char *fmt, va_list ap) {
         if (n > 0) {
           if (str_l < str_m) {
             size_t avail = str_m-str_l;
-            fast_memset(str+str_l, '0', (n>avail?avail:n));
+            fast_memset(str+str_l, '0', ((size_t)n>avail?avail:n));
           }
           str_l += n;
         }
@@ -988,7 +988,7 @@ int portable_vsnprintf(char *str, size_t str_m, const char *fmt, va_list ap) {
           if (str_l < str_m) {
             size_t avail = str_m-str_l;
             fast_memcpy(str+str_l, str_arg+zero_padding_insertion_ind,
-                        (n>avail?avail:n));
+                        ((size_t)n>avail?avail:n));
           }
           str_l += n;
         }
@@ -999,7 +999,7 @@ int portable_vsnprintf(char *str, size_t str_m, const char *fmt, va_list ap) {
         if (n > 0) {
           if (str_l < str_m) {
             size_t avail = str_m-str_l;
-            fast_memset(str+str_l, ' ', (n>avail?avail:n));
+            fast_memset(str+str_l, ' ', ((size_t)n>avail?avail:n));
           }
           str_l += n;
         }
